enel-sap-sp/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0
pyodbc>=5.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pywin32>=306; platform_system == "Windows"

numpy
